Re-added user-centered report. Unified dependency injection for readers.
